Corrections in live classifier and Addons to SentimentAnalysis  On branch live_classifier  Changes to be committed: 	modified:   Sentiment_Analysis.ipynb I have added a code to download the reported dataset in github I have added a code to extract the .tar.gz file downloaded I have modifed train/pos and train/neg paths since now any user can download the dataset and extract it to local repo I have included nltk downloads commands to download necessary files for processing.
	modified:   live_classifier.py
	new file:   pickled_algos/documents.pickle
	new file:   pickled_algos/word_features5k.pickle

Live_classifier has an error. it calls for an EnsembleClassifier function
but a VoteClassifier function is actually declared. I renamed the vote
function to ensemble to correct this.

Also this script asks for documents.pickle which is not in original repo
I have processed the jupyter notebook to get this file and
word_features5k

Also I added the function print_feat so that I can print the features
words. If I miss understood and those are not the features words which
I interpret as the most relevant words found so that algorithm can tell
if it is a positive or negative review, please remove that function